Update databases directly
Instead of storing SQL statements in /MAMP/htdocs/ folders and opening a browser window to execute a PHP script that executes those SQL statements on the local MySQL database; and using phpMyAdmin at the hosting site to apply those SQL statements to the remote database, log in the local and remotes databases and execute SQL satements directly.

In order to log in to the remote database, that database had to be moved from the InfinityFree hosting site to HelioHost.

Also, installed packages were brought up to date.